Set delay and sound timers 8 bit size

+++ src/registers.rs
@@ -3,8 +3,8 @@ pub const V_REGISTERS_SIZE: usize = 16;
 pub struct Registers {
     pub v: [u8; V_REGISTERS_SIZE],
     pub i: u16,
-    pub delay_timer: u16,
-    pub sound_timer: u16,
+    pub delay_timer: u8,
+    pub sound_timer: u8,
     pub program_counter: u16,
 }
